+++ CHANGELOG.md
@@ -1,9 +1,45 @@
 ## [Unreleased]
 
-## [0.1.6] - Initial stable release
-- Initial release
-## [0.1.7] - Updates for accessibility
-- Update all svg elements to have a role="img"
-- Update all svg elements to have a descriptive <title> element
-- Update all svg elements to have an aria-labelledby attribute referencing the <title> element
+## [0.2.9]
+- Bump gem version
 
+## [0.2.8]
+- Improve error message when icon is not found
+
+## [0.2.7]
+- Allow overriding the title attribute
+
+## [0.2.6]
+- Revert addition of title attribute for accessibility
+
+## [0.2.5]
+- Add support for title attribute and symbols for name and type arguments
+
+## [0.2.4]
+- Bump version
+
+## [0.2.3]
+- Add support for custom style attributes
+
+## [0.2.2]
+- Add support for custom data attributes
+
+## [0.2.1]
+- Bump version
+
+## [0.2.0]
+- Bump supported version of activesupport
+
+## [0.1.9]
+- Fix default width/height class overriding
+
+## [0.1.8]
+- Change required nokogiri version to 1.14.3
+
+## [0.1.7]
+- Update all SVG elements to have a role="img"
+- Update all SVG elements to have a descriptive `<title>` element
+- Update all SVG elements to have an aria-labelledby attribute referencing the `<title>` element
+
+## [0.1.6]
+- Initial stable release